mlops/project/fraud-detection-mlops/requirements.txt
# Core data processing
pandas==2.1.4
numpy==1.26.4
scipy==1.11.4

# Machine learning
scikit-learn==1.4.0
imbalanced-learn==0.11.0

# Deep learning
torch==2.2.0
transformers==4.37.2
datasets==2.16.1
tokenizers==0.15.1

# Embeddings
gensim==4.3.2

# Experiment tracking
mlflow==2.10.0
boto3==1.34.34

# API serving
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1

# Monitoring
prometheus-client==0.19.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.18.0

# Notebook
jupyter==1.0.0
ipykernel==6.29.2
nbformat==5.9.2

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
tqdm==4.66.1
joblib==1.3.2

# Testing
pytest==7.4.4
pytest-cov==4.1.0
httpx==0.26.0

# Code quality
flake8==7.0.0
black==24.1.1
